DRAFT: Add support for grpc-web
This is a basic implementation of the grpc-web protocol over HTTP/2.

Do NOT use this as-is in production; it does not properly configure
HTTP access controls through CORS.

Unfortunately, some of the changes here are a bit at odds with the
current class structure:
- The Deframer is wrapped, but the code expects it *not* to be wrapped
  in some places
- The Framer uses Sink which we overwrite with custom implementations
  of framing (to implement Base64 encoding) and writeTrailers

It also also unclear to me under what circumstances headers may not be
written yet when trailers are sent.

Also note that this *DOES NOT* add support for grpc-web over HTTP/1.1.

Related to https://github.com/grpc/grpc-java/issues/4823.