DRAFT: Allow access to the HTTP2 streams
This is a fairly hacky way of allowing gRPC and (normal) HTTP2 calls to
coexist on the same port. If an incoming HTTP2 call is *not* a gRPC
call, we simply forward it to an application-provided handler. This is
exclusive to netty. It also adds some API surface to the gRPC Java
library that may be undesirable from a maintenance point of view.

There are a few hacks here:
- Injection of the handler through a static field (this could be moved
  to the netty server builder)
- Extension of NettyServerStream.TransportState, and calls into private
  methods of the super class
- Reuse of the SendGrpcFrameCommand class with non-gRPC frames (I would
  argue that the class should be renamed, as its not at all specific to
  gRPC)

Another question is how flow control should be handled.

Related to https://github.com/grpc/grpc-java/issues/3458.